requirements.txt
fastapi
uvicorn
sqlalchemy
email-validator
psycopg2-binary
python-dotenv
passlib[bcrypt]